modified:   tcp-forwarder/src/pools.rs

+++ tcp-forwarder/src/pools.rs
@@ -428,7 +428,8 @@ async fn dynamic_filler_task(
         let min_needed = dynamic_config.min_size.saturating_sub(current_total);
 
         // 检查可用连接是否不足
-        let available_threshold = (current_total * dynamic_config.scaling.target_buffer_ratio).max(5); // 至少保持5个可用连接
+        let available_threshold =
+            ((current_total as f64) * dynamic_config.scaling.target_buffer_ratio).max(5.0) as usize; // 至少保持5个可用连接
         let availability_needed = available_threshold.saturating_sub(current_available);
 
         let needed = min_needed.max(availability_needed);
